lws-struct: json and sqlite on 32-bit OS
Test that we can handle int -1 correctly on sqlite and
greater than 32-bit unsigned on json with ull

+++ lib/misc/lws-struct-sqlite.c
@@ -74,13 +74,13 @@ lws_struct_sq3_deser_cb(void *priv, int cols, char **cv, char **cn)
 				if (map->aux == sizeof(int)) {
 					int *pi;
 					pi = (int *)(u + map->ofs);
-					*pi = atoi(cv[n]);
+					*pi = (int)atoll(cv[n]); /* 32-bit OS */
 					break;
 				}
 				if (map->aux == sizeof(long)) {
 					long *pl;
 					pl = (long *)(u + map->ofs);
-					*pl = atol(cv[n]);
+					*pl = (long)atoll(cv[n]); /* 32-bit OS */
 					break;
 				}
 				{
